refactor code from main to a separate function, create standalone preprocess.py with functions for calculating features
